Only perform an image compare if the pixel heights of the sources images are under 1000px different. Otherwise the compare takes too long.

+++ folderList.php
@@ -108,51 +108,66 @@ if($tier != 4) {
 	
 				// Check if the diff already exists and if not, create it
 				if(!file_exists($_SERVER['DOCUMENT_ROOT'] . "/" .$diffsPath."/".$fileName)) {
-	
-					// Doesn't exist - create the shell command
-					$compareCommand = "compare -metric RMSE -fuzz 8% -highlight-color Magenta -subimage-search ";
-					// Check the image pixel heights and if different, put the larger one first - required by 'compare'
-					if(getImageSize($fullPath1E.$fileName)[1] >= getImageSize($fullPath2E.$fileName)[1]) {
-						$compareCommand .= $fullPath1E.$fileName . " " . $fullPath2E.$fileName;
-						file_put_contents($logFile,"1E first\n",FILE_APPEND);
+						
+					// Check the file sizes and if they're too different, skip the comparison and log via image and text file
+					// Get the two image heights
+					$image1Height = getImageSize($fullPath1E.$fileName)[1];
+					$image2Height = getImageSize($fullPath2E.$fileName)[1];
+					
+					// Only run the comparison if the height difference between the two files is under 1000px
+					if(abs($image1Height-$image2Height) < 1000) {
+
+						// Doesn't exist and sizes are OK - create the shell command
+						$compareCommand = "compare -metric RMSE -fuzz 8% -highlight-color Magenta -subimage-search ";
+						
+						// Check the image pixel heights and if different, put the larger one first - required by 'compare'
+						if($image1Height >= $image2Height) {
+							$compareCommand .= $fullPath1E.$fileName . " " . $fullPath2E.$fileName;
+							file_put_contents($logFile,"1E first\n",FILE_APPEND);
+						} else {
+							$compareCommand .= $fullPath2E.$fileName . " " . $fullPath1E.$fileName;
+							file_put_contents($logFile,"2E first\n",FILE_APPEND);
+						}
+						$compareCommand .= " " . $_SERVER['DOCUMENT_ROOT'] . "/" .$diffsPath."/".$fileName;
+		
+						// Log the compare command
+						file_put_contents($logFile,date("Y-m-d H:i:s")." ".$fileName . ": Comparing via: ".$compareCommand."\n",FILE_APPEND);
+						
+						// Run the compare command
+						shell_exec($compareCommand) . "<br><br>";
+						
+						// Check for -0 and -1 files which are generated if they're different sizes and if they exist, delete -1 and rename -0 to the original name
+						$zeroName = substr($fileName, 0, strrpos($fileName, '.')) . "-0" . substr($fileName, strrpos($fileName, '.'));
+						$oneName = substr($fileName, 0, strrpos($fileName, '.')) . "-1" . substr($fileName, strrpos($fileName, '.'));
+						file_put_contents($logFile, "Zero name: " . $zeroName . "\n");
+						file_put_contents($logFile, "One name: " . $oneName . "\n");
+						
+						if(file_exists($diffsPath."/".$zeroName)) {
+							rename($diffsPath."/".$zeroName,$diffsPath."/".$fileName);
+						}
+						if(file_exists($diffsPath."/".$oneName)) {
+							unlink($diffsPath."/".$oneName);
+						}
+		
+						$diffOutput = shell_exec("convert " . $diffsPath."/".$fileName . " -fill black +opaque \"rgb(255,0,255)\" -format %c histogram:info:");
+						file_put_contents($logFile,date("Y-m-d H:i:s")."\n".$diffOutput."\n",FILE_APPEND);
+						
+						$diffOutput = explode("\n",$diffOutput);
+						$diffOutputFine = trim(substr($diffOutput[0],0,strpos($diffOutput[0],":")));
+						$diffOutputBroke = trim(substr($diffOutput[1],0,strpos($diffOutput[1],":")));
+						if($diffOutputBroke == "") { $diffOutputBroke = 0; };
+						
+						file_put_contents($logFile,date("Y-m-d H:i:s")." OK:".$diffOutputFine."|\n",FILE_APPEND);
+						file_put_contents($logFile,date("Y-m-d H:i:s")." Broke:".$diffOutputBroke."|\n",FILE_APPEND);
+						
+						$pcBroken = round($diffOutputBroke/($diffOutputBroke+$diffOutputFine)*100);
+						file_put_contents($diffsPath."/".$fileName.".txt","Percent of image mismatched based on area of magenta found:\n".$pcBroken."%\n");
 					} else {
-						$compareCommand .= $fullPath2E.$fileName . " " . $fullPath1E.$fileName;
-						file_put_contents($logFile,"2E first\n",FILE_APPEND);
+						// File sizes are too big to perform the diff
+						copy($_SERVER['DOCUMENT_ROOT'] . "/Error-SizesTooDifferent.png",$_SERVER['DOCUMENT_ROOT'] . "/" .$diffsPath."/".$fileName);
+						file_put_contents($diffsPath."/".$fileName.".txt","Percent of image mismatched based on area of magenta found:\nFile sizes too different to perform the compare\n");
+						file_put_contents($logFile,date("Y-m-d H:i:s"). $diffsPath."/".$fileName . ": notifications written in place of comparison being performed because pixel height differential of source files was too big"."\n",FILE_APPEND);
 					}
-					$compareCommand .= " " . $_SERVER['DOCUMENT_ROOT'] . "/" .$diffsPath."/".$fileName;
-	
-					// Log the compare command
-					file_put_contents($logFile,date("Y-m-d H:i:s")." ".$fileName . ": Comparing via: ".$compareCommand."\n",FILE_APPEND);
-					
-					// Run the compare command
-					shell_exec($compareCommand) . "<br><br>";
-					
-					// Check for -0 and -1 files which are generated if they're different sizes and if they exist, delete -1 and rename -0 to the original name
-					$zeroName = substr($fileName, 0, strrpos($fileName, '.')) . "-0" . substr($fileName, strrpos($fileName, '.'));
-					$oneName = substr($fileName, 0, strrpos($fileName, '.')) . "-1" . substr($fileName, strrpos($fileName, '.'));
-					file_put_contents($logFile, "Zero name: " . $zeroName . "\n");
-					file_put_contents($logFile, "One name: " . $oneName . "\n");
-					
-					if(file_exists($diffsPath."/".$zeroName)) {
-						rename($diffsPath."/".$zeroName,$diffsPath."/".$fileName);
-					}
-					if(file_exists($diffsPath."/".$oneName)) {
-						unlink($diffsPath."/".$oneName);
-					}
-	
-					$diffOutput = shell_exec("convert " . $diffsPath."/".$fileName . " -fill black +opaque \"rgb(255,0,255)\" -format %c histogram:info:");
-					file_put_contents($logFile,date("Y-m-d H:i:s")."\n".$diffOutput."\n",FILE_APPEND);
-					
-					$diffOutput = explode("\n",$diffOutput);
-					$diffOutputFine = trim(substr($diffOutput[0],0,strpos($diffOutput[0],":")));
-					$diffOutputBroke = trim(substr($diffOutput[1],0,strpos($diffOutput[1],":")));
-					if($diffOutputBroke == "") { $diffOutputBroke = 0; };
-					
-					file_put_contents($logFile,date("Y-m-d H:i:s")." OK:".$diffOutputFine."|\n",FILE_APPEND);
-					file_put_contents($logFile,date("Y-m-d H:i:s")." Broke:".$diffOutputBroke."|\n",FILE_APPEND);
-					
-					$pcBroken = round($diffOutputBroke/($diffOutputBroke+$diffOutputFine)*100);
-					file_put_contents($diffsPath."/".$fileName.".txt","Percent of image mismatched based on area of magenta found:\n".$pcBroken."%\n",FILE_APPEND);
 					
 				} else {
 					file_put_contents($logFile,date("Y-m-d H:i:s")." ".$fileName . ": diff already generated - no compare needed\n",FILE_APPEND);
